backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl==3.1.2
reportlab==4.0.9
pillow==10.2.0
drf-spectacular==0.27.1

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
